imp: adding 08-wasm build opts with libwasmvm linking disabled (#5923)
* wip: messing with build options for wasm cgo

* chore: mv type assertion to wasm_cgo with build flags

* chore: mv make target to build section

* chore: revert cgo enabled 0 build opt for testing

* chore: rm unneeded file

* update build tag

* linter

* update panic message

* add Codec back

* refactor: adapt build tags to match wasmvm

* Update modules/light-clients/08-wasm/doc.go


* Update modules/light-clients/08-wasm/keeper/keeper_vm.go


* chore: make lint-fix

* chore: make lint-fix

---------
