+++ modules/light-clients/08-wasm/internal/ibcwasm/wasm_vm.go
@@ -1,0 +1,7 @@
+//go:build cgo && !nolink_libwasmvm
+
+package ibcwasm
+
+import wasmvm "github.com/CosmWasm/wasmvm/v2"
+
+var _ WasmEngine = (*wasmvm.VM)(nil)
